Fix List of Hparams
Fix a bug where concrete list of hparams required an extra level of dictionary. For example, given a dataclass of:

```python
@dataclass
class Foo(hp.Hparams):
    baz: int = hp.required(doc='int')

@dataclass
class ParentListHPNoRegistry(hp.Hparams):
    foos: List[Foo] = hp.required(doc='All the foos')
```

The following YAML should be allowed:

```yaml
foos:
  - baz: 2
  - baz: 3
```

Before, one had to specify a phantom key in the yaml, like this:

```yaml
foos:
  foo+1:
    - baz: 2
  foo+2:
     - baz: 3
```

The old syntax is still allowed, but now raises a deprecation warning. This is a non-breaking change.

Closes https://mosaicml.atlassian.net/browse/CO-766